Fix/dialog stop timer confirmation (#6648)
* feat: create a dialog for exit confirmation

* feat: create a dialog for logout confirmation

* fix: auth guard

* fix: unecessary relation to optimize query

* feat: show exit confirmation dialog before to quit if app running

* feat: show logout confirmation dialog before to logout if app is running

* fix: update timer current session on start accordingly to remote timer's current session.

* feat: add blur transition to image viewer

* Update dialog-stop-timer-exit-confirmation.ts

* Update dialog-stop-timer-logout-confirmation.ts

* feat: add spinner until image show

* fix: dialog detail message

* fix: revert removed code.

---------

+++ packages/desktop-ui-lib/src/lib/settings/settings.component.ts
@@ -670,8 +670,8 @@ export class SettingsComponent implements OnInit, AfterViewInit {
 		});
 
 		this.electronService.ipcRenderer.on('_logout_quit_install_', (event, arg) => {
-			this._ngZone.run(() => {
-				this.logout(true);
+			this._ngZone.run(async () => {
+				await this.restartAndUpdate();
 			});
 		});
 	}
@@ -787,11 +787,11 @@ export class SettingsComponent implements OnInit, AfterViewInit {
 		this.updateSetting(value, 'prerelease');
 	}
 
-	restartApp() {
+	public async restartApp(): Promise<void> {
 		if (this.isServer && this.serverIsRunning) {
 			this._restartDisable$.next(true);
 		} else {
-			this.logout();
+			await this.logout();
 		}
 		const thConfig = {};
 		this.thirdPartyConfig.forEach((item) => {
@@ -852,8 +852,8 @@ export class SettingsComponent implements OnInit, AfterViewInit {
 		this.electronService.ipcRenderer.send('check_for_update');
 	}
 
-	restartAndUpdate() {
-		this.logout(true);
+	public async restartAndUpdate(): Promise<void> {
+		await this.logout(true);
 	}
 
 	toggleAwView(value) {
@@ -897,10 +897,26 @@ export class SettingsComponent implements OnInit, AfterViewInit {
 	/*
 	 * Logout desktop timer
 	 */
-	logout(isAfterUpgrade?: boolean) {
-		console.log('On Logout');
-		localStorage.clear();
-		this.electronService.ipcRenderer.send('logout_desktop', isAfterUpgrade);
+	public async logout(isAfterUpgrade?: boolean): Promise<void> {
+		let isLogout = true;
+
+		if (this.appSetting?.timerStarted) {
+			isLogout = await this.electronService.ipcRenderer.invoke(
+				'LOGOUT_STOP'
+			);
+		} else {
+			localStorage.clear();
+		}
+
+		if (isLogout) {
+			console.log('On Logout');
+			this.electronService.ipcRenderer.send(
+				'logout_desktop',
+				isAfterUpgrade
+			);
+		} else {
+			console.log('Logout Cancelled');
+		}
 	}
 
 	onServerChange(val) {
